Renamed files to match rest of repo. Moved sources to src/ dir.
