+++ CMakeLists.txt
@@ -102,7 +102,7 @@ generate_messages(
 ## CATKIN_DEPENDS: catkin_packages dependent projects also need
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
-   INCLUDE_DIRS include
+#  INCLUDE_DIRS include
 #  LIBRARIES robosub
   CATKIN_DEPENDS roscpp rospy std_msgs message_runtime
 #  DEPENDS system_lib
@@ -132,7 +132,6 @@ catkin_package(
 include_directories(
   ${catkin_INCLUDE_DIRS}
   src
-  include
 )
 
 #macro for adding and linking nodes to the ros framework
